% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_txshift.R
\name{aipw_txshift}
\alias{aipw_txshift}
\title{Compute Augmented Inverse Probability Weighted Estimate of Counterfactual
Mean Under Shifted Treatment}
\usage{
aipw_txshift(data_internal, C = rep(1, nrow(data_internal)), V = NULL,
  delta, cens_weights, cens_weights_norm, ipcw_estim, Qn_estim, Hn_estim,
  fluc_method = c("standard", "weighted"),
  eif_tol = 1/nrow(data_internal), max_iter = 10000,
  eif_reg_type = c("hal", "glm"), ipcw_fit_args, ipcw_fit_type,
  ipcw_efficiency = TRUE)
}
\arguments{
\item{data_internal}{...}

\item{C}{A \code{numeric} binary vector giving information on whether a given
observation was subject to censoring. This is used to compute an IPCW-TMLE
in cases where two-stage sampling is performed. The default assumes that no
censoring was present (i.e., a two-stage design was NOT used). N.B., this is
equivalent to the term %\Delta in the notation used in the original Rose and
van der Laan manuscript that introduced/formulated IPCW-TML estimators.}

\item{V}{The covariates that are used in determining the sampling procedure
that gives rise to censoring. The default is \code{NULL} and corresponds to
scenarios in which there is no censoring (in which case all values in the
preceding argument \code{C} must be uniquely 1. To specify this, pass in a
NAMED \code{list} identifying variables amongst W, A, Y that are thought to
have played a role in defining the sampling/censoring mechanism (C).}

\item{delta}{A \code{numeric} value indicating the shift in the treatment to
be used in defining the target parameter. This is defined with respect to
the scale of the treatment (A).}

\item{cens_weights}{...}

\item{cens_weights_norm}{...}

\item{ipcw_estim}{...}

\item{Qn_estim}{...}

\item{Hn_estim}{...}

\item{fluc_method}{The method to be used in submodel fluctuation step of
the TMLE computation. The choices are "standard" and "weighted".}

\item{eif_tol}{A \code{numeric} giving the convergence criterion for the TML
estimator. This is the the maximum mean of the efficient influence function
(EIF) to be used in declaring convergence (theoretically, should be zero).}

\item{max_iter}{A \code{numeric} integer giving the maximum number of steps
to be taken in iterating to a solution of the efficient influence function.}

\item{eif_reg_type}{Whether a flexible nonparametric function ought to be
used in the dimension-reduced nuisance regression of the targeting step for
the censored data case. By default, the method used is a nonparametric
regression based on the Highly Adaptive Lasso (from package \code{hal9001}).
Set this to \code{"glm"} to instead use a simple linear regression model.
In this step, the efficient influence function (EIF) is regressed against
covariates contributing to the censoring mechanism (i.e., EIF ~ V | C = 1).}

\item{ipcw_fit_args}{...}

\item{ipcw_fit_type}{...}

\item{ipcw_efficiency}{Whether to invoke an augmentation of the IPCW-TMLE
procedure that performs an iterative process to ensure efficiency of the
resulting estimate. The default is \code{TRUE}; only set to \code{FALSE} if
possible inefficiency of the IPCW-TMLE is not a concern.}
}
\value{
S3 object of class \code{txshift} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
Compute Augmented Inverse Probability Weighted Estimate of Counterfactual
Mean Under Shifted Treatment
}
