% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_propensity}
\alias{bound_propensity}
\title{Bound Generalized Propensity Score (Conditional Density)}
\usage{
bound_propensity(vals)
}
\arguments{
\item{vals}{\code{numeric} vector of propensity score estimate values. Note
that, for this parameter, the propensity score is (conditional) density and
so it ought not be bounded from above.}
}
\value{
TODO
}
\description{
Bound Generalized Propensity Score (Conditional Density)
}
\details{
TODO
}
