% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_scaling}
\alias{bound_scaling}
\title{Scaling by Inducing Boundedness}
\usage{
bound_scaling(Y, pred_vals = NULL, scale_target = Y,
  scale_type = c("bound_in_01", "observed_vals"))
}
\arguments{
\item{Y}{A \code{numeric} vector corresponding to the observed values of the
outcome variable of interest.}

\item{pred_vals}{A \code{numeric} vector corresponding to predicted values of
the outcome of interest (i.e., Qn in the Targeted Learning notation).}

\item{scale_target}{A \code{numeric} vector specifying the quantity that is
to be re-scaled in the way specified below in \code{scale_type}.}

\item{scale_type}{An atomic \code{character} vector specifying the type of
scaling to be performed. Use "bound_in_01" to force \code{scale_target}
above to be bounded in the interval (0, 1) with respect to the outcome
\code{Y}. The other option, "observed_vals", re-scales \code{scale_target}
to be on the same scale as the input \code{Y}.}
}
\description{
description
}
\keyword{internal}
