% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci_shifttx}
\alias{ci_shifttx}
\title{Confidence Intervals for Shifted Treatment Parameters}
\usage{
ci_shifttx(tmle_shift, level = 0.95)
}
\arguments{
\item{tmle_shift}{An tmle_shift of class \code{shifttx}, as produced by invoking
the function \code{tmle_shifttx}, for which a confidence interval is to be
computed.}

\item{level}{A \code{numeric} indicating the level of the confidence interval
to be computed.}
}
\description{
description
}
