% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.txshift}
\alias{confint.txshift}
\title{Confidence Intervals for Shifted Treatment Parameters}
\usage{
\method{confint}{txshift}(object, parm = seq_len(object$psi), level = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{txshift}, as produced by invoking
the function \code{tmle_txshift}, for which a confidence interval is to be
computed.}

\item{parm}{A \code{numeric} vector indicating indices of \code{object$est}
for which to return confidence intervals.}

\item{level}{A \code{numeric} indicating the level of the confidence interval
to be computed.}

\item{...}{Other arguments. Not currently used.}
}
\description{
Compute confidence intervals for estimates produced by \code{tmle_txshift}
}
