% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eifs.R
\name{eif}
\alias{eif}
\title{Compute the Shift Parameter Estimate and the Efficient Influence Function}
\usage{
eif(Y, Qn, Hn, estimator = c("tmle", "onestep"), fluc_mod_out = NULL,
  Delta = rep(1, length(Y)), ipc_weights = rep(1, length(Y)),
  ipc_weights_norm = rep(1, length(Y)), eif_tol = 1e-07)
}
\arguments{
\item{Y}{A \code{numeric} vector of the observed outcomes.}

\item{Qn}{An object providing the value of the outcome evaluated after
imposing a shift in the treatment. This object is passed in after being
constructed by a call to the internal function \code{est_Q}.}

\item{Hn}{An object providing values of the auxiliary ("clever") covariate,
constructed from the treatment mechanism and required for targeted minimum
loss-based estimation. This object object should be passed in after being
constructed by a call to the internal function \code{est_Hn}.}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood estimation or \code{"onestep"} for a one-step
augmented inverse probability weighted (AIPW) estimator.}

\item{fluc_mod_out}{An object giving values of the logistic regression fit to
traverse the submodel for targeted minimum loss-based estimation. This type
of object should be the output of the internal routines to perform this
step of the TMLE computation, as given by \code{fit_fluctuation}.}

\item{Delta}{Indicator for missingness. Used for compatibility with the
routine to compute IPCW-TML estimates.}

\item{ipc_weights}{A \code{numeric} vector that gives inverse probability of
censoring weights for each observation. These are generated by invoking the
routines for estimating the censoring mechanism.}

\item{eif_tol}{A \code{numeric} giving the minimum value of the tolerance to
be used in evaluating whether the computation of the efficient influence
function has converged.}
}
\description{
Computes the value of the treatment shift parameter as well as statistical
inference for the parameter based on the efficient influence function of the
target parameter, which takes the following form:
%D(P)(o) = H(a,w)(y - \bar{Q}(a,w)) + \bar{Q}(d(a,w)) - \psi(P)
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
