% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_shift.R
\name{est.equation}
\alias{est.equation}
\title{Estimating equation}
\usage{
est.equation(eps)
}
\arguments{
\item{eps}{Residuals from previous...}

\item{gn0d}{}
}
\description{
Estimating equation
}
\author{
Ivan Diaz, with contributions from Nima Hejazi
}
