% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_shift.R
\name{est_eqn}
\alias{est_eqn}
\title{Estimating equation}
\usage{
est_eqn(eps, Y, A, W, delta, QnAW, Qn, H1, gn0d, EQnd, D2, prev_sum)
}
\arguments{
\item{eps}{...}

\item{Y}{...}

\item{A}{...}

\item{W}{...}

\item{delta}{...}

\item{QnAW}{...}

\item{Qn}{...}

\item{H1}{...}

\item{gn0d}{...}

\item{EQnd}{...}

\item{D2}{...}

\item{prev_sum}{...}
}
\description{
Estimating equation
}
\author{
Ivan Diaz

Nima Hejazi
}
