% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-auxiliary_covar.R
\name{est_Hn}
\alias{est_Hn}
\title{Estimate the Auxiliary ("Clever") Covariate}
\usage{
est_Hn(gn, a = NULL, w = NULL)
}
\arguments{
\item{gn}{...}

\item{a}{...}

\item{w}{...}
}
\description{
description
}
\keyword{internal}
