% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_shift.R
\name{f.iter}
\alias{f.iter}
\title{Iterative function}
\usage{
f.iter(Qn, gn, gn0d = NULL, prev.sum = 0, first = FALSE)
}
\arguments{
\item{Qn}{Function to compute Q(A, W) = E(Y | A, W).}

\item{gn}{Function to compute g(A, W) = density(A | W).}

\item{gn0d}{Same as the function gn, modified from the previous iteration.}

\item{prev.sum}{...}

\item{first}{...}

\item{delta}{}
}
\description{
Iterative function
}
\author{
Ivan Diaz, with contributions from Nima Hejazi
}
