% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-density_hal.R
\name{format_long_hazards}
\alias{format_long_hazards}
\title{Generate long format hazards data for conditional density estimation}
\usage{
format_long_hazards(A, W, n_bins = 10)
}
\arguments{
\item{A}{...}

\item{W}{...}

\item{n_bins}{...}
}
\description{
description forthcoming
}
