% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-eif_ipcw.R
\name{ipcwtmle_eif}
\alias{ipcwtmle_eif}
\title{Compute IPCW Part of the Efficient Influence Function for IPCW-TMLE}
\usage{
ipcwtmle_eif(fluc_fit_out, Hn, Y, C, ipc_weights = rep(1, length(Y)),
  tol_eif = 1e-07)
}
\arguments{
\item{fluc_fit_out}{...}

\item{Hn}{...}

\item{Y}{A \code{numeric} vector of observed outcomes.}

\item{ipc_weights}{...}

\item{tol_eif}{...}
}
\description{
Computes an additional component of the efficient influence function needed
for efficient estimation of IPCW-TMLEs. This takes the form:
%0=P_n(\Delta-\Pi_n^*(V)))\frac{\E(D^F(P^0_{X,n})\mid\Delta=1,V)}{\Pi^*_n(V)}
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
