% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm_vim_shift.R
\name{msm_shift}
\alias{msm_shift}
\title{Test for a trend in the causal effects of a grid of shift interventions}
\usage{
msm_shift(tmle_fit_estimates, delta_grid, level = 0.95, weights = NULL,
  ci_type = c("simultaneous", "marginal"))
}
\arguments{
\item{tmle_fit_estimates}{A \code{list} corresponding to the \code{tml_est}
slot of an object of class \code{txshift_vim}, containing TML estimates over
a grid of posited shift interventions produced from \code{tmle_txshift}.}

\item{delta_grid}{A \code{numeric} vector giving the individual values of the
shift parameter used in computing each of the TML estimates.}

\item{level}{The nominal coverage probability of the confidence interval.}

\item{weights}{A \code{numeric} vector indicating the weights (if any) to be
applied to each of the estimated mean counterfactual outcomes under posited
values of the shift parameter delta. The default is to weight each estimate
by the inverse of its variance, in order to improve stability; however, this
may be changed depending on the exact choice of shift function.}
}
\description{
Test for a trend in the causal effects of a grid of shift interventions
}
