% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm_vim.R
\name{msm_vimshift}
\alias{msm_vimshift}
\title{Marginal structural model for the causal effects of a grid of interventions}
\usage{
msm_vimshift(Y, A, W, C = rep(1, length(Y)), V = NULL,
  delta_grid = seq(-0.5, 0.5, 0.5), estimator = c("tmle", "onestep"),
  weighting = c("variance", "identity"), ci_level = 0.95,
  ci_type = c("simultaneous", "marginal"), ...)
}
\arguments{
\item{Y}{A \code{numeric} vector of the observed outcomes.}

\item{A}{A \code{numeric} vector corresponding to a treatment variable. The
parameter of interest is defined as a location shift of this quantity.}

\item{W}{A \code{matrix}, \code{data.frame}, or similar corresponding to a
set of baseline covariates.}

\item{C}{A \code{numeric} binary vector giving information on whether a given
observation was subject to censoring. This is used to compute an IPCW-TMLE
in cases where two-stage sampling is performed. The default assumes that no
censoring was present (i.e., a two-stage design was NOT used). N.B., this is
equivalent to the term %\Delta in the notation used in the original Rose and
van der Laan manuscript that introduced/formulated IPCW-TML estimators.}

\item{V}{The covariates that are used in determining the sampling procedure
that gives rise to censoring. The default is \code{NULL} and corresponds to
scenarios in which there is no censoring (in which case all values in the
preceding argument \code{C} must be uniquely 1. To specify this, pass in a
NAMED \code{list} identifying variables amongst W, A, Y that are thought to
have played a role in defining the sampling/censoring mechanism (C).}

\item{delta_grid}{A \code{numeric} vector giving the individual values of the
shift parameter used in computing each of the TML estimates.}

\item{estimator}{The type of estimator to be fit, either \code{"tmle"} for
targeted maximum likelihood estimation or \code{"onestep"} for a one-step
augmented inverse probability weighted (AIPW) estimator.}

\item{weighting}{Whether to weight each parameter estimate by the inverse of
its variance (in order to improve stability of the resultant MSM fit) or to
simply weight all parameter estimates equally.}

\item{ci_level}{A \code{numeric} indicating the desired coverage level of the
confidence interval to be computed.}

\item{ci_type}{Whether to construct a simultaneous confidence band covering
all parameter estimates at once or marginal confidence intervals covering
each parameter estimate separately.}

\item{...}{Additional arguments to be passed to \code{txshift}.}
}
\description{
Marginal structural model for the causal effects of a grid of interventions
}
\examples{
set.seed(429153)
n_obs <- 100
W <- as.numeric(replicate(1, rbinom(n_obs, 1, 0.5)))
A <- as.numeric(rnorm(n_obs, mean = 2 * W, sd = 1))
Y <- rbinom(n_obs, 1, plogis(A + W + rnorm(n_obs, mean = 0, sd = 1)))
msm <- msm_vimshift(
  W = W, A = A, Y = Y, estimator = "tmle",
  g_fit_args = list(
    fit_type = "hal",
    n_bins = 5,
    grid_type = "equal_mass",
    lambda_seq = exp(seq(-1, -9,
      length = 100
    ))
  ),
  Q_fit_args = list(
    fit_type = "glm",
    glm_formula = "Y ~ ."
  )
)
}
