% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nan_to_num}
\alias{nan_to_num}
\title{Map NaN Values to A Chosen Value}
\usage{
nan_to_num(nums_in, map_val = 0)
}
\arguments{
\item{nums_in}{A \code{numeric} vector containing some number of values that
are represented as \code{NaN}.}
}
\description{
This is a heinous transformation function that makes it possible to map
values of \code{NaN} to zeros. This may be useful for bookkeeping purposes
but it is recommended that you never use this. It appears that this also
works when when \code{NaN} is substituted for \code{NA}.
}
\keyword{internal}
