% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.shifttx}
\alias{print.shifttx}
\title{Print Method for shifttx Objects}
\usage{
\method{print}{shifttx}(x, ...)
}
\arguments{
\item{x}{An object of class \code{shifttx}.}

\item{...}{Other options (not currently used).}
}
\description{
The \code{print} method for objects of class \code{shifttx}.
}
