% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.txshift}
\alias{print.txshift}
\title{Print Method for txshift Objects}
\usage{
\method{print}{txshift}(x, ...)
}
\arguments{
\item{x}{An object of class \code{txshift}.}

\item{...}{Other options (not currently used).}
}
\description{
The \code{print} method for objects of class \code{txshift}.
}
