% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_methods.R
\name{shift_additive}
\alias{shift_additive}
\title{Simple Additive Modified Treatment Policy}
\usage{
shift_additive(A, W = NULL, delta)
}
\arguments{
\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{A \code{numeric} indicating the magnitude of the shift to be
computed for the treatment \code{A}.}
}
\description{
Simple Additive Modified Treatment Policy
}
\keyword{internal}
