% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target-qn_mech.R
\name{target_qn}
\alias{target_qn}
\title{Targeting for Part of EIF Corresponding to Weighted Censoring Distribution}
\usage{
target_qn(Qn_shift, ipc_weights, data_in)
}
\arguments{
\item{Qn_shift}{...}

\item{ipc_weights}{...}

\item{data_in}{...}
}
\description{
DESCRIPTION HERE
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
