% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_eif.R
\name{tmle_eif}
\alias{tmle_eif}
\title{Compute the Shift Parameter Estimate and the Efficient Influence Function}
\usage{
tmle_eif(fluc_fit_out, Hn, Y, tol_eif = 1e-07)
}
\arguments{
\item{fluc_fit_out}{...}

\item{Hn}{...}

\item{Y}{A \code{numeric} vector of observed outcomes.}

\item{tol_eif}{...}
}
\description{
Computes the value of the treatment shift parameter as well as statistical
inference for the parameter based on the efficient influence function of the
target parameter, which takes the following form:
%D(P)(o) = H(a,w)(y - \bar{Q}(a,w)) + \bar{Q}(d(a,w)) - \psi(P)
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
