% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_tmle_ID.R
\name{tmle_shift_original_ID}
\alias{tmle_shift_original_ID}
\title{TML Estimate of the Effect of a Continuous Treatment}
\usage{
tmle_shift_original_ID(Y, A, W, Qn, gn, delta, tol = 1e-05,
  iter_max = 5, A_val)
}
\arguments{
\item{Y}{A \code{numeric} vector of observed outcomes.}

\item{A}{A \code{numeric} vector of observed treatments.}

\item{W}{A \code{matrix} or \code{data.frame} of baseline covariates.}

\item{Qn}{Function to compute the outcome regression: Q(A, W) = E(Y | A, W).}

\item{gn}{Function to compute the propensity score: g(A, W) = density(A | W).}

\item{delta}{A \code{numeric} for the shift to be placed on the treatment of
interest (i.e., the effect of shifting treatment \code{A} by delta units).}

\item{tol}{A \code{numeric} for the tolerance for measuring convergence of
the parametric fluctuation.}

\item{iter_max}{A \code{numeric} for the maximum number of iterations.}

\item{A_val}{A \code{vector} of \code{numeric} values for the points in the
range of the treatment \code{A} to approximate integrals by Riemmann sums.
Note that these must be equally spaced along a grid.}
}
\description{
TML Estimate of the Effect of a Continuous Treatment
}
\examples{
\dontrun{
n <- 100
W <- data.frame(W1 = runif(n), W2 = rbinom(n, 1, 0.7))
A <- rpois(n, lambda = exp(3 + 0.3 * log(W$W1) - 0.2 * exp(W$W1) * W$W2))
Y <- rbinom(n, 1, plogis(-1 + 0.05 * A - 0.02 * A * W$W2 +
  0.2 * A * tan(W$W1^2) - 0.02 * W$W1 * W$W2 +
  0.1 * A * W$W1 * W$W2))
fitA.0 <- glm(A ~ I(log(W1)) + I(exp(W1)):W2,
  family = poisson,
  data = data.frame(A, W)
)
fitY.0 <- glm(Y ~ A + A:W2 + A:I(tan(W1^2)) + W1:W2 + A:W1:W2,
  family = binomial, data = data.frame(A, W)
)
gn.0 <- function(A = A, W = W) {
  dpois(A, lambda = predict(fitA.0, newdata = W, type = "response"))
}
Qn.0 <- function(A = A, W = W) {
  predict(fitY.0,
    newdata = data.frame(A, W, row.names = NULL),
    type = "response"
  )
}
tmle00 <- tmle_shift_original_ID(
  Y = Y, A = A, W = W, Qn = Qn.0, gn = gn.0, delta = 2,
  tol = 1e-4, iter_max = 5, A_val = seq(1, 60, 1)
)
}

}
\author{
Iván Díaz
}
\keyword{internal}
