% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_shifttx.R
\name{tmle_shifttx}
\alias{tmle_shifttx}
\title{Compute Targeted Minimum Loss-Based Estimates of Treatment Shift Parameters}
\usage{
tmle_shifttx(W, A, Y, delta, parallel)
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Y}{...}

\item{delta}{...}

\item{parallel}{...}
}
\value{
S3 object of class \code{shifttx} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
description THIS IS A USER-FACING WRAPPER FUNCTION
}
