% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txshift.R
\name{tx_shift}
\alias{tx_shift}
\title{Simple Modified Treatment Policy (Shifted Treatment)}
\usage{
tx_shift(A, W = NULL, delta)
}
\arguments{
\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{A \code{numeric} indicating the magnitude of the shift to be
computed for the treatment \code{A}.}
}
\description{
Simple Modified Treatment Policy (Shifted Treatment)
}
\keyword{internal}
