% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_scaling}
\alias{bound_scaling}
\title{Scaling by Inducing Boundedness}
\usage{
bound_scaling(Y, scaled_vals, rescale_to = c("quasibinomial",
  "observed"))
}
\arguments{
\item{Y}{A \code{numeric} vector corresponding to the observed values of the
outcome variable of interest.}

\item{scaled_vals}{A \code{numeric} vector specifying the quantity that is
to be re-scaled in the way specified below in \code{rescale_to}.}

\item{rescale_to}{An atomic \code{character} vector specifying the type of
scaling to be performed. Use "quasibinomial" to force \code{scaled_vals}
above to be bounded in the interval (0, 1) with respect to the outcome
\code{Y}. The other option, "observed", re-scales \code{scaled_vals}
to be on the same original scale as the input \code{Y}.}
}
\description{
description
}
\keyword{internal}
