% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bound_scaling}
\alias{bound_scaling}
\title{Induce Scaling}
\usage{
bound_scaling(Y, preds_scaled = NULL, scale = c("zero_one", "original"))
}
\arguments{
\item{Y}{...}

\item{preds_scaled}{...}

\item{scale}{...}
}
\description{
description
}
\keyword{internal}
