% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.shifttx}
\alias{confint.shifttx}
\title{Confidence Intervals for Shifted Treatment Parameters}
\usage{
\method{confint}{shifttx}(tmle_shifttx_out, level = 0.95)
}
\description{
Confidence Intervals for Shifted Treatment Parameters
}
