% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.shifttx}
\alias{confint.shifttx}
\title{Confidence Intervals for Shifted Treatment Parameters}
\usage{
\method{confint}{shifttx}(object, parm = seq_len(object$psi), level = 0.95,
  ...)
}
\arguments{
\item{object}{An object of class \code{shifttx}, as produced by invoking
the function \code{tmle_shifttx}, for which a confidence interval is to be
computed.}

\item{parm}{A \code{numeric} vector indicating indices of \code{object$est}
for which to return confidence intervals.}

\item{level}{A \code{numeric} indicating the level of the confidence interval
to be computed.}

\item{...}{Other arguments. Not currently used.}
}
\description{
description
}
