% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-propensity_score.R
\name{est_g}
\alias{est_g}
\title{Estimate the Treatment Mechanism (Propensity Score)}
\usage{
est_g(A, W, delta = 0, ipc_weights = rep(1, length(A)),
  fit_type = c("sl", "glm"), sl_lrnrs_dens = NULL,
  std_args = list(nbins = 25, bin_method = "dhist", bin_estimator =
  condensier::speedglmR6$new(), parfit = FALSE))
}
\arguments{
\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{A \code{numeric} value identifying a shift in the observed value
of the treatment under which observations are to be evaluated.}

\item{ipc_weights}{A \code{numeric} vector of observation-level weights, as
produced by the internal procedure to estimate the censoring mechanism
\code{estimate-ipc_weights}.}

\item{fit_type}{A \code{character} specifying whether to use Super Learner to
fit the density estimation.}

\item{sl_lrnrs_dens}{An object containing a set of instantiated learners from
the \code{sl3} package, to be used in fitting an ensemble model.}

\item{std_args}{A \code{list} of arguments to be passed to \code{fit_density}
from the \code{condensier} package when the argument \code{fit_type} is set
to "glm" (not invoking Super Learner).}
}
\description{
Compute the propensity score (treatment mechanism) for the observed data,
including the shift. This returns the propensity score for the observed data
(at A_i) and the shift of interest (at A_i - delta).
}
