% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-propensity_score.R
\name{est_g}
\alias{est_g}
\title{Estimate the Treatment Mechanism}
\usage{
est_g(A, W, delta = 0, ...)
}
\arguments{
\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{A \code{numeric} ...}

\item{...}{}
}
\description{
Compute the propensity score (treatment mechanism) for the observed data,
including the shift. This returns the propensity score for the observed data
(at A_i) and the shift of interest (at A_i - delta).
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
