% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-auxiliary_covar.R
\name{est_Hn}
\alias{est_Hn}
\title{Estimate the Auxiliary ("Clever") Covariate}
\usage{
est_Hn(gn, a = NULL, w = NULL)
}
\arguments{
\item{gn}{An estimate of the treatment probability (propensity score), using
the output provided by internal function \code{estimate-propensity_score}.}

\item{a}{A \code{numeric} vector of the observed values of the treatment.}

\item{w}{A \code{numeric}, \code{matrix}, \code{data.frame} or similar object
that contains the observed values of the baseline covariates.}
}
\description{
Estimate the Auxiliary ("Clever") Covariate
}
\keyword{internal}
