% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mechanisms.R
\name{est_Hn}
\alias{est_Hn}
\title{Estimate Auxiliary Covariate from Efficient Influence Function}
\usage{
est_Hn(gn)
}
\arguments{
\item{gn}{An estimate of the treatment probability (propensity score), using
the output provided by internal function \code{estimate-propensity_score}.}
}
\value{
A \code{data.table} with two columns, containing estimates of the
 auxiliary covariate at the natural value of the exposure H(A, W) and at the
 shifted value of the exposure H(A + delta, W).
}
\description{
Estimate Auxiliary Covariate from Efficient Influence Function
}
\details{
Compute an estimate of the auxiliary covariate required to update
 initial estimates via logistic tilting models in targeted minimum loss
 estimation.
}
