% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-ipc_weights.R
\name{est_ipcw}
\alias{est_ipcw}
\title{Estimate Inverse Probability of Censoring Weights}
\usage{
est_ipcw(V, Delta)
}
\arguments{
\item{V}{...}

\item{Delta}{...}
}
\value{
A \code{numeric} vector corresponding to the inverse probability of
 censoring weights required for computing an IPCW-TMLE. Formally, this is
 nothing more than %\frac{\Delta}{\Pi_n}, where the term %\Pi_n is simply the
 predicted probability of belonging to a censoring class as computed using
 standard logistic regression.
}
\description{
description
}
