% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-ipc_weights.R
\name{est_ipcw}
\alias{est_ipcw}
\title{Estimate Inverse Probability of Censoring Weights}
\usage{
est_ipcw(V, Delta, fit_type = c("sl", "glm"), glm_formula = "Delta ~ .",
  sl_lrnrs = NULL)
}
\arguments{
\item{V}{...}

\item{Delta}{...}

\item{fit_type}{A \code{character} indicating whether to perform the fit
using GLMs or a Super Learner. If use of Super Learner is desired, then the
argument \code{sl_lrnrs} must be provided.}

\item{glm_formula}{...}

\item{sl_lrnrs}{An \code{R6} object of class \code{Lrnr_sl}, a Super Learner
object created externally using the \code{sl3} package.}
}
\value{
A \code{numeric} vector corresponding to the inverse probability of
 censoring weights required for computing an IPCW-TMLE. Formally, this is
 nothing more than %\frac{\Delta}{\Pi_n}, where the term %\Pi_n is simply the
 predicted probability of belonging to a censoring class as computed using
 standard logistic regression.
}
\description{
description
}
