% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-outcome_reg.R
\name{est_Q}
\alias{est_Q}
\title{Estimate the Outcome Regression}
\usage{
est_Q(Y, A, W, delta = 0, glm_form = "Y ~ .", sl_lrnrs = NULL,
  sl_meta = NULL)
}
\arguments{
\item{Y}{A \code{numeric} vector of observed outcomes.}

\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{...}

\item{glm_form}{...}

\item{sl_meta}{...}

\item{sl_learners}{...}
}
\description{
Compute the outcome regression for the observed data, including with the
shift imposed by the intervention. This returns the propensity score for the
observed data (at A_i) and the shift of interest (at A_i - delta).
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
