% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_shift.R
\name{f_iter}
\alias{f_iter}
\title{Fluctuation (iterative)}
\usage{
f_iter(Qn, gn, gn0d = NULL, prev_sum = 0, first = FALSE, h_int, W, A,
  A_val, Y, delta)
}
\arguments{
\item{Qn}{Function to compute Q(A, W) = E(Y | A, W).}

\item{gn}{Function to compute g(A, W) = density(A | W).}

\item{gn0d}{\code{Numeric} computed for gn from the first iteration only.}

\item{prev_sum}{...}

\item{first}{...}

\item{h_int}{...}

\item{W}{...}

\item{A}{...}

\item{A_val}{...}

\item{Y}{...}

\item{delta}{...}
}
\description{
Fluctuation (iterative)
}
\author{
Ivan Diaz

Nima Hejazi
}
