% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuate-submodel_reg.R
\name{fit_fluc}
\alias{fit_fluc}
\title{Fit Logistic Regression to Traverse the Fluctuation Submodel}
\usage{
fit_fluc(Y, Qn_scaled, Hn, ipc_weights = rep(1, length(Y)),
  method = c("standard", "weighted"))
}
\arguments{
\item{Y}{...}

\item{Qn_scaled}{...}

\item{Hn}{...}

\item{ipc_weights}{...}

\item{method}{...}
}
\description{
description
}
\keyword{internal}
