% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-fluctuate_reg.R
\name{fit_fluc}
\alias{fit_fluc}
\title{Fit Logistic Regression in the Fluctuation Submodel}
\usage{
fit_fluc(Y, Qn_scaled, Hn, method = c("standard", "weighted"))
}
\arguments{
\item{Y}{...}

\item{Qn_scaled}{...}

\item{Hn}{...}

\item{method}{...}
}
\description{
description
}
\keyword{internal}
