% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuate-submodel_reg.R
\name{fit_fluc}
\alias{fit_fluc}
\title{Fit Logistic Regression to Traverse the Fluctuation Submodel}
\usage{
fit_fluc(Y, Qn_scaled, Hn, ipc_weights = rep(1, length(Y)),
  method = c("standard", "weighted"))
}
\arguments{
\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{Qn_scaled}{...}

\item{Hn}{...}

\item{ipc_weights}{A \code{numeric} vector that gives inverse probability of
censoring weights for each observation. These are generated by invoking the
routines for estimating the censoring mechanism.}

\item{method}{A \code{character} giving the type of regression to be used in
traversing the fluctuation sub-model. Choices are "weighted" and "standard".
Please consult the literature for details on the differences.}
}
\description{
description
}
\keyword{internal}
