% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_txshift.R
\name{fit_fluctuation}
\alias{fit_fluctuation}
\title{Fit Logistic Regression to Traverse the Fluctuation Submodel}
\usage{
fit_fluctuation(Y, Qn_scaled, Hn, ipc_weights = rep(1, length(Y)),
  method = c("standard", "weighted"), flucmod_tol = 100)
}
\arguments{
\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{Qn_scaled}{An object providing the value of the outcome evaluated after
imposing a shift in the treatment. This object should be passed in after
being constructed by a call to the internal function \code{est_Q}.}

\item{Hn}{An object providing values of the auxiliary ("clever") covariate,
constructed from the treatment mechanism and required for targeted minimum
loss-based estimation. This object object should be passed in after being
constructed by a call to the internal function \code{est_Hn}.}

\item{ipc_weights}{A \code{numeric} vector that gives inverse probability of
censoring weights for each observation. These are generated by invoking the
routines for estimating the censoring mechanism.}

\item{method}{A \code{character} giving the type of regression to be used in
traversing the fluctuation sub-model. Choices are "weighted" and "standard".
Please consult the literature for details on the differences.}

\item{flucmod_tol}{A \code{numeric} indicating the largest value to be
tolerated in the fluctuation model for the targeted minimum loss estimator.}
}
\description{
Fit Logistic Regression to Traverse the Fluctuation Submodel
}
