% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-density_hal.R
\name{format_long_hazards}
\alias{format_long_hazards}
\title{Generate long format hazards data for conditional density estimation}
\usage{
format_long_hazards(A, W, wts = rep(1, length(A)),
  type = c("equal_range", "equal_mass", "equal_width"), n_bins = 10,
  width = NULL)
}
\arguments{
\item{A}{...}

\item{W}{...}

\item{wts}{...}

\item{type}{...}

\item{n_bins}{...}

\item{width}{...}
}
\description{
Generate long format hazards data for conditional density estimation
}
