% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_original}
\alias{scale_to_original}
\title{Scale transformed values in the unit interval to the original scale}
\usage{
scale_to_original(scaled_vals, max_orig, min_orig)
}
\arguments{
\item{scaled_vals}{A \code{numeric} vector corresponding to re-scaled values
in the unit interval, to be re-scaled to the original interval.}

\item{max_orig}{A \code{numeric} scalar value giving the maximum of the
values on the original scale.}

\item{min_orig}{A \code{numeric} scalar value giving the minimum of the
values on the original scale.}
}
\description{
Scale transformed values in the unit interval to the original scale
}
\keyword{internal}
