% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_unit}
\alias{scale_to_unit}
\title{Transform values by scaling to the unit interval}
\usage{
scale_to_unit(vals)
}
\arguments{
\item{vals}{A \code{numeric} vector corresponding to the observed values of
the variable of interest, to be re-scaled to the unit interval [0,1].}
}
\value{
A \code{numeric} vector of the same length as \code{vals}, where the
 values are re-scaled to lie in unit interval [0, 1].
}
\description{
Transform values by scaling to the unit interval
}
\details{
A transformation that scales an arbitrary set of input values to
 the unit interval. See \code{\link{scale_to_original}} for a corresponding
 backtransformation.
}
