% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_funs.R
\name{shift_additive}
\alias{shift_additive}
\title{Simple Additive Modified Treatment Policy}
\usage{
shift_additive(A, W = NULL, delta)
}
\arguments{
\item{A}{A \code{numeric} vector of observed treatment values.}

\item{W}{A \code{numeric} matrix of observed baseline covariate values.}

\item{delta}{A \code{numeric} indicating the magnitude of the shift to be
computed for the treatment \code{A}.}
}
\value{
A \code{numeric} vector containing the shifted exposure values.
}
\description{
Simple Additive Modified Treatment Policy
}
\details{
A simple modified treatment policy that modifes the observed value
 of the exposure by shifting it by a value \code{delta}. Note that this
 shifting function assumes support of A|W across all strata of W.
}
