% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.txshift}
\alias{summary.txshift}
\title{Summary for Shifted Treatment Parameter Objects}
\usage{
\method{summary}{txshift}(object, ..., ci_level = 0.95)
}
\arguments{
\item{object}{An object of class \code{txshift}, as produced by invoking
the function \code{tmle_txshift}, for which a confidence interval is to be
computed.}

\item{...}{Other arguments. Not currently used.}

\item{ci_level}{A \code{numeric} indicating the level of the confidence
interval to be computed.}
}
\description{
Print a convenient summary for objects computed using \code{tmle_txshift}.
}
