% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-eif_std.R
\name{tmle_eif}
\alias{tmle_eif}
\title{Compute the Shift Parameter Estimate and the Efficient Influence Function}
\usage{
tmle_eif(fluc_mod_out, Hn, Y, Delta = rep(1, length(Y)),
  ipc_weights = rep(1, length(Y)), ipc_weights_norm = rep(1, length(Y)),
  tol_eif = 1e-07)
}
\arguments{
\item{fluc_mod_out}{...}

\item{Hn}{...}

\item{Y}{A \code{numeric} vector of observed outcomes.}

\item{Delta}{Indicator for missingness. Used for compatibility with the
routine to compute IPCW-TML estimates.}

\item{ipc_weights}{...}

\item{tol_eif}{...}
}
\description{
Computes the value of the treatment shift parameter as well as statistical
inference for the parameter based on the efficient influence function of the
target parameter, which takes the following form:
%D(P)(o) = H(a,w)(y - \bar{Q}(a,w)) + \bar{Q}(d(a,w)) - \psi(P)
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
