% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-eif_ipcw.R
\name{tmle_eif_ipcw}
\alias{tmle_eif_ipcw}
\title{Compute IPC-Weighted Component of Efficient Influence Function}
\usage{
tmle_eif_ipcw(fluc_mod_out, data_in, Hn, Delta, ipc_weights = rep(1,
  nrow(data_in)), ipc_weights_norm = rep(1, nrow(data_in)), eif_tol = 1e-10)
}
\arguments{
\item{fluc_mod_out}{...}

\item{data_in}{...}

\item{Hn}{...}

\item{Delta}{...}

\item{ipc_weights}{...}

\item{ipc_weights_norm}{...}

\item{eif_tol}{...}
}
\description{
Computes an additional component of the efficient influence function needed
for efficient estimation of IPCW-TMLEs. This takes the form:
%0=P_n(\Delta-\Pi_n^*(V)))\frac{\E(D^F(P^0_{X,n})\mid\Delta=1,V)}{\Pi^*_n(V)}
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
