% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-eif_ipcw.R
\name{tmle_eif_ipcw}
\alias{tmle_eif_ipcw}
\title{Compute IPC-Weighted Component of Efficient Influence Function}
\usage{
tmle_eif_ipcw(fluc_mod_out, data_in, Hn, Delta, ipc_weights = rep(1,
  nrow(data_in)), ipc_weights_norm = rep(1, nrow(data_in)), eif_tol = 1e-10)
}
\arguments{
\item{fluc_mod_out}{An object giving values of the logistic regression fit to
traverse the submodel for targeted minimum loss-based estimation. This type
of object should be the output of the internal routines to perform this
step of the TMLE computation, as given by \code{fit_fluc}.}

\item{data_in}{A \code{data.frame} or similar object containing the observed
data structure as produced by early steps of \code{tmle_txshift}.}

\item{Hn}{An object providing values of the auxiliary ("clever") covariate,
constructed from the treatment mechanism and required for targeted minimum
loss-based estimation. This object object should be passed in after being
constructed by a call to the internal function \code{est_Hn}.}

\item{Delta}{Indicator for missingness. Used for compatibility with the
routine to compute IPCW-TML estimates.}

\item{ipc_weights}{A \code{numeric} vector that gives inverse probability of
censoring weights for each observation. These are generated by invoking the
routines for estimating the censoring mechanism.}

\item{ipc_weights_norm}{A \code{numeric} vector of the same weights described
in the previous argument, except normalized in this case.}

\item{eif_tol}{A \code{numeric} giving the minimum value of the tolerance to
be used in evaluating whether the computation of the efficient influence
function has converged.}
}
\description{
Computes an additional component of the efficient influence function needed
for efficient estimation of IPCW-TMLEs. This takes the form:
%0=P_n(\Delta-\Pi_n^*(V)))\frac{\E(D^F(P^0_{X,n})\mid\Delta=1,V)}{\Pi^*_n(V)}
}
\author{
Nima Hejazi

David Benkeser
}
\keyword{internal}
