% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-shifttx.R
\name{tmle_shifttx}
\alias{tmle_shifttx}
\title{Compute Targeted Maximum Likelihood Estimate for Treatment Shift Parameter}
\usage{
tmle_shifttx(W, A, Y, C = rep(1, length(Y)), delta = 0,
  fit_type = c("glm", "sl"), fluc_method = c("standard", "weighted"),
  eif_tol = 1e-07, args = list(ipcw_fit = list(glm_formula = "Delta ~ .",
  sl_lrnrs = NULL), g_fit = list(nbins = 25, bin_method = "dhist", bin_estimator
  = condensier::speedglmR6$new(), parfit = FALSE, sl_lrnrs_dens = NULL), Q_fit =
  list(glm_formula = "Y ~ .", sl_lrnrs = NULL)))
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Y}{...}

\item{C}{A \code{numeric} binary vector giving information on whether a given
observation was subject to censoring. This is used to compute an IPCW-TMLE
in cases where two-stage sampling is performed. The default assumes that no
censoring was present (i.e., a two-stage design was NOT used). N.B., this is
equivalent to the term %\Delta in the notation used in the original Rose and
van der Laan manuscript that introduced/formulated IPCW-TML estimators.}

\item{delta}{...}

\item{fit_type}{...}

\item{fluc_method}{The method to be used in submodel fluctuation step of
the TMLE computation. The choices are "standard" and "weighted".}

\item{eif_tol}{The convergence criterion for the TML estimator. This is the
the maximum mean of the efficient influence function (EIF) to be accepted in
calling convergence (theoretically, the value is zero).}

\item{args}{A nested \code{list} of \code{list}s that specifies arguments to
be passed to the various internal functions for the estimation procedure.
Each of sub-list corresponds to a single internal function. As such, for
details on (1) \code{ipcw_fit}, see the documention of \code{est_ipcw}; (2)
\code{g_fit}, see the documentation of \code{est_g}; (3) \code{Q_fit}, see
the documentation for \code{est_Q}.}
}
\value{
S3 object of class \code{shifttx} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
description THIS IS A USER-FACING WRAPPER FUNCTION
}
