% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-shifttx.R
\name{tmle_shifttx}
\alias{tmle_shifttx}
\title{Compute Targeted Maximum Likelihood Estimate for Treatment Shift Parameter}
\usage{
tmle_shifttx(W, A, Y, delta, C = rep(1, length(Y)),
  ipcw_fit_args = list(fit_type = "glm", glm_formula = "Delta ~ .", sl_lrnrs =
  NULL, sl_task = NULL), g_fit_args = list(nbins = 20, bin_method = "dhist",
  bin_estimator = condensier::speedglmR6$new(), parfit = FALSE),
  Q_fit_args = list(fit_method = "glm", glm_formula = "Y ~ .", sl_lrnrs =
  NULL), fluc_method = "standard", eif_tol = 1e-07)
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Y}{...}

\item{delta}{...}

\item{C}{A \code{numeric} binary vector giving information on whether a given
observation was subject to censoring. This is used to compute an IPCW-TMLE
in cases where two-stage sampling is performed. The default assumes that no
censoring was present (i.e., a two-stage design was NOT used). N.B., this is
equivalent to the term %\Delta in the notation used in the original Rose and
van der Laan manuscript that introduced/formulated IPCW-TML estimators.}

\item{g_fit_args}{A \code{list} of arguments to be passed to the internal
function \code{est_g}, which then passes these same arguments on to the
function \code{fit_density} of the package \code{condensier}.}

\item{Q_fit_args}{A \code{list} of arugments to be passed to the function to
estimate the outcome regression. Consult the documentation for the function
\code{est_Q} for details.}

\item{fluc_method}{...}

\item{eif_tol}{...}
}
\value{
S3 object of class \code{shifttx} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
description THIS IS A USER-FACING WRAPPER FUNCTION
}
