% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-shifttx.R
\name{tmle_shifttx}
\alias{tmle_shifttx}
\title{Compute Targeted Minimum Loss-Based Estimates of Treatment Shift Parameters}
\usage{
tmle_shifttx(W, A, Y, delta, parallel = NULL, Q_fit_method = c("glm", "sl"),
  fluc_method = c("standard", "weighted"), eif_tol = 1e-07, ...)
}
\arguments{
\item{W}{...}

\item{A}{...}

\item{Y}{...}

\item{delta}{...}

\item{parallel}{...}

\item{Q_fit_method}{...}

\item{fluc_method}{...}

\item{eif_tol}{...}

\item{...}{Passed to internal function \code{est_g}, which then internally
passes these arguments to \code{fit_density} from package \code{condensier}.}
}
\value{
S3 object of class \code{shifttx} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
description THIS IS A USER-FACING WRAPPER FUNCTION
}
