% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle-txshift.R
\name{tmle_txshift}
\alias{tmle_txshift}
\title{Compute Targeted Maximum Likelihood Estimate for Treatment Shift Parameter}
\usage{
tmle_txshift(W, A, Y, C = rep(1, length(Y)), V = NULL, delta = 0,
  fluc_method = c("standard", "weighted"), eif_tol = 1e-09,
  max_iter = 10000, ipcw_fit_args = list(fit_type = c("glm", "sl"),
  glm_formula = "Delta ~ .", sl_lrnrs = NULL), g_fit_args = list(fit_type =
  c("glm", "sl"), nbins = 35, bin_method = "dhist", bin_estimator =
  condensier::speedglmR6$new(), parfit = FALSE, sl_lrnrs_dens = NULL),
  Q_fit_args = list(fit_type = c("glm", "sl"), glm_formula = "Y ~ .", sl_lrnrs
  = NULL))
}
\arguments{
\item{W}{A \code{matrix} or \code{data.frame} corresponding to a set of
baseline covariates.}

\item{A}{A \code{numeric} vector corresponding to a treatment variable. The
parameter of interest is defined as a location shift of this quantity.}

\item{Y}{A \code{numeric} vector corresponding to an outcome variable.}

\item{C}{A \code{numeric} binary vector giving information on whether a given
observation was subject to censoring. This is used to compute an IPCW-TMLE
in cases where two-stage sampling is performed. The default assumes that no
censoring was present (i.e., a two-stage design was NOT used). N.B., this is
equivalent to the term %\Delta in the notation used in the original Rose and
van der Laan manuscript that introduced/formulated IPCW-TML estimators.}

\item{V}{The covariates that are used in determining the sampling procedure
that gives rise to censoring. The default is \code{NULL} and corresponds to
scenarios in which there is no censoring (in which case all values in the
preceding argument \code{C} must be uniquely 1. To specify this, pass in a
NAMED \code{list} identifying variables amongst W, A, Y that are thought to
have played a role in defining the sampling/censoring mechanism (C).}

\item{delta}{A \code{numeric} value indicating the shift in the treatment to
be used in defining the target parameter. This is defined with respect to
the scale of the treatment (A).}

\item{fluc_method}{The method to be used in submodel fluctuation step of
the TMLE computation. The choices are "standard" and "weighted".}

\item{eif_tol}{A \code{numeric} giving the convergence criterion for the TML
estimator. This is the the maximum mean of the efficient influence function
(EIF) to be used in declaring convergence (theoretically, should be zero).}

\item{max_iter}{A \code{numeric} integer giving the maximum number of steps
to be taken in iterating to a solution of the efficient influence function.}

\item{ipcw_fit_args}{...}

\item{g_fit_args}{...}

\item{Q_fit_args}{...}
}
\value{
S3 object of class \code{txshift} containing the results of the
 procedure to compute a TML estimate of the treatment shift parameter.
}
\description{
description THIS IS A USER-FACING WRAPPER FUNCTION
}
