% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interv_shift.R
\name{tx_shift}
\alias{tx_shift}
\title{Shift Treatments}
\usage{
tx_shift(A, W = NULL, delta, type = "additive", direc = c("down", "up"))
}
\arguments{
\item{A}{...}

\item{W}{...}

\item{delta}{...}

\item{type}{...}

\item{direc}{...}
}
\description{
description
}
\keyword{internal}
